add SetAttrOnFile/GetAttrFromFile stub impl.
OSX does not expose the immutable flag for files via ioctl().
